fix: add `.cargo-checksum.json` at root of fetched Git repos
This only works if the crate is at the root of the repo
and not in a workspace.

We'll need to use the `cargo` lib to handle all that complexity,
namely:

a) figuring out the path to the root of the workspace from the
   root of the repo (they may not always coincide, e.g. if the
   Rust code lives under a `rust/` directory in the repo;

b) figuring out the path from the root of the workspace to the
   package we're looking for;

c) replace all workspace-inherited keys in the target package's
   Cargo.toml (see https://github.com/rust-lang/cargo/issues/11192);

And to handle that we'll need to have proper callback-enabled
lazy-chaining, instead of the current eval-string based approach.